Add base seeder and development seeder
Change necessary to have different seed procedures for
different environments. For development, the seed is
always applied with basic habits and user. For staging and
production only the User is seeded.
